requirements.txt
dash~=2.17.1
pandas~=2.2.2
vaderSentiment~=3.3.2
plotly~=5.22.0
asyncio~=3.4.3
ccxt~=4.4.14
pytz~=2024.1
networkx~=3.3
numpy~=1.26.4
statsmodels~=0.14.2
PyYAML~=6.0.1
python-dateutil~=2.9.0.post0
streamlit~=1.36.0
setuptools~=70.3.0
krakenex~=2.2.2
scikit-learn~=1.5.2
seaborn~=0.13.2
matplotlib~=3.9.2
dtreeviz~=2.2.2